Added new config values

+++ common/src/main/java/de/mrjulsen/paw/event/ModCommonEvents.java
@@ -5,6 +5,7 @@ import java.util.Optional;
 
 import org.joml.Vector3f;
 
+import de.mrjulsen.paw.config.ModCommonConfig;
 import de.mrjulsen.paw.mixin.DimensionDataStorageAccessor;
 import de.mrjulsen.wires.WiresApi;
 import de.mrjulsen.wires.graph.WireGraph;
@@ -23,12 +24,14 @@ public final class ModCommonEvents {
         LifecycleEvent.SERVER_LEVEL_LOAD.register((level) -> {
             if (level.isClientSide()) return;
             WireGraphManager.build(level);
-            try {
-                CompoundTag nbt = level.getDataStorage().readTagFromDisk("wiresapi_wire_network", 0);
-                WireGraphManager.get(level, WiresApi.PAW_CATENARY_WIRES).upgrade(nbt.getCompound("data"));
-                level.getDataStorage().save();
-                ((DimensionDataStorageAccessor)level.getDataStorage()).paw$getDataFile("wiresapi_wire_network").delete();
-            } catch (IOException e) {}
+            if (ModCommonConfig.CONVERT_WIRE_NETWORK.get()) {
+                try {
+                    CompoundTag nbt = level.getDataStorage().readTagFromDisk("wiresapi_wire_network", 0);
+                    WireGraphManager.get(level, WiresApi.PAW_CATENARY_WIRES).upgrade(nbt.getCompound("data"));
+                    level.getDataStorage().save();
+                    ((DimensionDataStorageAccessor)level.getDataStorage()).paw$getDataFile("wiresapi_wire_network").delete();
+                } catch (IOException e) {}
+            }
         });
 
         LifecycleEvent.SERVER_STOPPED.register((server) -> {
